PcSpeaker: fixed pause, merged InternalSpeaker class
